Put Dialog components into own files
