+++ moderatorFrontend/src/main/kotlin/webcore/BasicCustomContentDialog.kt
@@ -1,0 +1,34 @@
+package webcore
+
+import react.ChildrenBuilder
+import react.Props
+import react.State
+
+/**
+ * Convenience function for defining customContent in DialogConfig
+ * Can only be used for trivial cases where no extra props, state handling or callbacks are required other than a basic ChildrenBuilder block
+ * WILL NOT WORK FOR STATE UPDATES IN THE CALL SITE!
+ */
+fun basicCustomContent(block: ChildrenBuilder.() -> Unit) =
+  DialogConfig.CustomContent(component = BasicCustomContentDialog::class, setProps = {
+    config = BasicCustomContentDialogConfig(block = {
+      block()
+    })
+  })
+
+class BasicCustomContentDialogConfig(
+  val block: ChildrenBuilder.() -> Unit
+)
+
+external interface BasicCustomContentDialogProps : Props {
+  var config: BasicCustomContentDialogConfig
+}
+
+@Suppress("UPPER_BOUND_VIOLATED")
+class BasicCustomContentDialog(props: BasicCustomContentDialogProps) :
+  RComponentWithCoroutineScope<BasicCustomContentDialogProps, State>(props) {
+
+  override fun ChildrenBuilder.render() {
+    props.config.block(this)
+  }
+}
